requirements.txt
fastapi
uvicorn
python-dotenv
pydantic

langchain
langchain-community
langchain-core

faiss-cpu
pypdf

openai
google-generativeai
groq
boto3

streamlit
tqdm